requirements_uv.txt
absl-py==2.3.1
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-cors==0.8.1
aiosignal==1.4.0
airportsdata==20250909
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.75.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
appdirs==1.4.4
astor==0.8.1
asttokens==3.0.1
attrs==25.4.0
av==16.0.1
bitsandbytes==0.49.0
blake3==1.0.8
blobfile==3.0.0
build==1.3.0
cachetools==6.2.4
cbor2==5.7.1
certifi==2026.2.25
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.5
click==8.3.1
cloudpickle==3.1.2
codetiming==1.4.0
colorful==0.5.8
compressed-tensors==0.11.0
cuda-bindings==13.1.1
cuda-pathfinder==1.3.3
cuda-python==13.1.1
cupy-cuda12x==13.6.0
datasets==4.7.0
debugpy==1.8.0
decorator==5.2.1
decord==0.6.0
deepspeed==0.18.3
depyf==0.19.0
dill==0.4.0
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
dnspython==2.8.0
docstring-parser==0.17.0
einops==0.8.1
email-validator==2.3.0
evalplus==0.3.1
executing==2.2.1
fastapi==0.127.0
fastapi-cli==0.0.20
fastapi-cloud-cli==0.7.0
fastar==0.8.0
fastrlock==0.8.3
filelock==3.25.2
fire==0.7.1
flash-attn==2.8.3
flashinfer-python==0.3.1
frozendict==2.4.7
frozenlist==1.8.0
fsspec==2024.9.0
gguf==0.17.1
gitdb==4.0.12
gitpython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.193.0
google-auth==2.45.0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
hf-transfer==0.1.9
hf-xet==1.4.2
hjson==3.1.0
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.34.0
hydra-core==1.3.2
identify==2.6.15
idna==3.11
importlib-metadata==8.7.1
iniconfig==2.3.0
interegular==0.3.3
ipython==9.8.0
ipython-pygments-lexers==1.1.1
isort==7.0.0
jedi==0.19.2
jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
jsonlines==4.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lark==1.2.2
latex2sympy2-extended==1.11.0
liger-kernel==0.6.4
lightning-utilities==0.15.2
llguidance==0.7.30
llvmlite==0.44.0
lm-format-enforcer==0.11.3
loguru==0.7.3
loralib==0.1.2
lxml==6.0.2
markdown==3.10
markdown-it-py==4.0.0
markupsafe==3.0.3
math-verify==0.9.0
mathruler==0.1.0
matplotlib-inline==0.2.1
mdurl==0.1.2
mistral-common==1.8.8
modelscope==1.33.0
mpmath==1.3.0
msgpack==1.1.2
msgspec==0.20.0
multidict==6.7.1
multipledispatch==1.0.0
multiprocess==0.70.18
nest-asyncio==1.6.0
networkx==3.6.1
ninja==1.13.0
nltk==3.9.2
nodeenv==1.10.0
numba==0.61.2
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cudnn-frontend==1.17.0
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==13.590.44
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
openai==1.99.1
openai-harmony==0.0.4
opencensus==0.11.4
opencensus-context==0.1.3
opencv-fixer==0.2.5
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
opentelemetry-api==1.39.1
opentelemetry-exporter-prometheus==0.60b1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
optimum==2.1.0
optree==0.18.0
orjson==3.11.5
outlines==0.1.11
outlines-core==0.2.11
packaging==26.0
pandas==3.0.1
parso==0.8.5
partial-json-parser==0.2.1.1.post7
peft==0.18.0
pexpect==4.9.0
pillow==12.0.0
pip==25.3
platformdirs==4.5.1
pluggy==1.6.0
pre-commit==4.5.1
prometheus-client==0.23.1
prometheus-fastapi-instrumentator==7.1.0
prompt-toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.0
protobuf==5.29.6
psutil==7.1.3
ptyprocess==0.7.0
pure-eval==0.2.3
py-cpuinfo==9.0.0
py-spy==0.4.1
pyarrow==23.0.1
pyasn1==0.6.1
pyasn1-modules==0.4.2
pybase64==1.4.3
pybind11==3.0.1
pycountry==24.6.1
pycparser==2.23
pycryptodomex==3.23.0
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-extra-types==2.10.6
pydantic-settings==2.12.0
pygments==2.19.2
pylatexenc==2.10
pynvml==13.0.1
pyparsing==3.3.2
pyproject-hooks==1.2.0
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-multipart==0.0.21
pytz==2025.2
pyvers==0.1.0
pyyaml==6.0.3
pyzmq==27.1.0
qwen-vl-utils==0.0.14
ray==2.48.0
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
rich==14.3.3
rich-toolkit==0.17.1
rignore==0.7.6
rouge-score==0.1.2
rpds-py==0.30.0
rsa==4.9.1
ruff==0.14.10
safetensors==0.7.0
scipy==1.17.1
sentencepiece==0.2.1
sentry-sdk==2.48.0
setproctitle==1.3.7
setuptools==79.0.1
sgl-kernel==0.3.9.post2
sglang==0.5.2
shellingham==1.5.4
six==1.17.0
smart-open==7.5.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soxr==1.0.0
stack-data==0.6.3
starlette==0.50.0
stop-sequencer==1.2.3
sympy==1.14.0
tabulate==0.10.0
tempdir==0.7.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensordict==0.10.0
termcolor==3.3.0
tiktoken==0.12.0
timm==1.0.16
tokenizers==0.21.4
torch==2.8.0
torch-memory-saver==0.0.8
torchao==0.9.0
torchaudio==2.8.0
torchdata==0.11.0
torchmetrics==1.8.2
torchvision==0.23.0
tqdm==4.67.3
traitlets==5.14.3
transformers==4.55.2
transformers-stream-generator==0.0.5
tree-sitter==0.25.2
tree-sitter-python==0.25.0
triton==3.4.0
typer==0.24.1
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.40.0
uvloop==0.21.0
verl==0.7.0
virtualenv==20.35.4
vllm==0.11.0
wandb==0.23.1
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
werkzeug==3.1.4
wget==3.2
wheel==0.45.1
wrapt==2.0.1
xformers==0.0.32.post1
xgrammar==0.1.25
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.0
